apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pocketcoach-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
    "google-genai>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
